eme: move src/core/eme/handle_session_events.ts to src/core/eme/session_events_listener.ts as `SessionEventsListener is the new name of the exported function
